add session model
add filters and session model

+++ PSY-master/app/Http/Controllers/TrainingController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Log;
 use App\Models\Match;
 use App\Models\Psychologist;
 use App\Models\Visit;
@@ -9,10 +9,25 @@ use App\Models\Institute;
 use App\Models\Training;
 use App\Models\Shapah;
 use Illuminate\Support\Facades\Auth;
+use App\Models\Session;
+use Input;
 
 class TrainingController extends Controller {
 	public function index() {
-       $trainings = Training::all();
+            $trainings= new Training;
+            if ( Input::has('filter_guid') && trim(Input::get('filter_guid')) !== '' ){
+                $trainings = $trainings->where('guide_id', '=', Input::get('filter_guid'));
+            }
+            if ( Input::has('filter_guided') && trim(Input::get('filter_guided')) !== '' ){
+                $trainings = $trainings->where('guided_id', '=', Input::get('filter_guided'));
+            }
+            if (Input::has('filter_year') && trim(Input::get('filter_year')) !== ''){
+                $trainings = $trainings->where('training_year', '=', (string)Input::get('filter_year'));
+            }
+            if (Input::has('filter_hours') && trim(Input::get('filter_hours')) !== ''){
+                $trainings = $trainings->where('training_hours ', '=', Input::get('filter_hours'));
+            }
+            $trainings = $trainings->get();
 		return view( 'indexes.training', compact( 'trainings' ));
 	}
 
@@ -25,18 +40,30 @@ class TrainingController extends Controller {
 	}
         
         public function update( Training $training ) {
-		$input_data = $this->getFormUserData();
+		$input_data = Input::All();
 		$training->fill( $input_data );
 		$training->save();
 		//$this->setUserPermission( $psychologist );
 		return redirect()->route( 'training.index' );
 	}
         
-        public function edit( Training $training ) {
+        
+        public function edit($train ) {
+                $training= new Training;
+                $training=$training->where('id', '=', $train)->first();
+                if (!($training)){
+                     Log::info("NO training");
+                    $trainings = $trainings->get();
+                }
 		$is_new = false;
                 $psychologists = $this->getShapahPsychologists( Auth::user() );
 		return view( 'forms.new_training', compact( 'training', 'is_new','psychologists' ) );
 	}
+        /*public function edit( Training $training ) {
+		$is_new = false;
+                $psychologists = $this->getShapahPsychologists( Auth::user() );
+		return view( 'forms.new_training', compact( 'training', 'is_new','psychologists' ) );
+	}*/
         
         public function store() {
 		$user_data = \Request::all();
@@ -44,11 +71,39 @@ class TrainingController extends Controller {
                 $training->save();
 		return redirect()->route( 'training.index' );
 	}
-
-        public function destroy(Training $train) {
-		$train->delete();
-		return redirect()->route( 'training.index' );
+        
+        
+        public function destroy($training ) {
+                $session_num= new Session;
+                $session_num=$session_num->where('trining_id', '=',$training)->count();
+                if ($session_num > 0){
+                    $error="שגיאה: להדרכה זאת קיים דיווח מפגש הדרכה";
+                    Log::info("Training have sessions");
+                }
+                else {
+                    $train= new Training;
+                    $train=$train->where('id', '=', $training)->first();
+                    $train->delete();
+                }
+                $trainings= new Training;
+                $trainings= $trainings->get();
+                return view( 'indexes.training', compact( 'trainings' ,'error') );
+		//return redirect()->route( 'training.index' );
 	}
+        
+        
+        /*public function destroy(Training $train) {
+            $input = Input::all();
+                //$id=Input::has('delete_id');
+                Log::info("111111111111111111:");
+                Log::info("t:".$train);
+                Log::info($train->id);
+                Log::info(gettype ($train->id));
+                Log::info($input);
+                \DB::table( 'Trainings' )->where('id', '=', $train->id)->delete();
+                $e=$train->delete();
+		return redirect()->route( 'training.index' );
+	}*/
         
  
          private function getShapahPsychologists( Psychologist $psychologist ) {
